embedder/requirements.txt
openai
supabase>=2.28.0
python-dotenv
